# swagger.yaml
openapi: 3.0.0
info:
  title: DIGIPIN API
  description: Encode and decode DIGIPIN based on latitude and longitude
  version: 1.0.0
servers:
  - url: http://localhost:5000/api

paths:
  /digipin/encode:
    post:
      summary: Encode lat/lon into DIGIPIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        "200":
          description: Successfully encoded DIGIPIN
          content:
            application/json:
              schema:
                type: object
                properties:
                  digipin:
                    type: string
        "400":
          description: Invalid input
    get:
      summary: Encode latitude and longitude into DIGIPIN
      parameters:
        - in: query
          name: latitude
          required: true
          schema:
            type: number
        - in: query
          name: longitude
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Successfully encoded DIGIPIN
          content:
            application/json:
              schema:
                type: object
                properties:
                  digipin:
                    type: string
        "400":
          description: Invalid input

  /digipin/decode:
    post:
      summary: Decode DIGIPIN to coordinates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                digipin:
                  type: string
      responses:
        "200":
          description: Successfully decoded DIGIPIN
          content:
            application/json:
              schema:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
        "400":
          description: Invalid DIGIPIN
    get:
      summary: Decode DIGIPIN to coordinates
      parameters:
        - in: query
          name: digipin
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully decoded DIGIPIN
          content:
            application/json:
              schema:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
        "400":
          description: Invalid DIGIPIN

  /digipin/validate:
    get:
      summary: Validate a DIGIPIN
      description: Checks whether a given DIGIPIN is valid based on its format and characters.
      parameters:
        - in: query
          name: digipin
          required: true
          schema:
            type: string
            example: "FJ4M-925C-7T"
      responses:
        "200":
          description: DIGIPIN validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean
                    example: true
        "400":
          description: Invalid request or DIGIPIN format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid DIGIPIN"